+++ src/stdlib/net_tcp.rs
@@ -147,7 +147,8 @@ impl Interp {
                 // AScript can read `listener.port` — makes `listen(host, 0)` usable.
                 let port = listener.local_addr().map(|a| a.port()).unwrap_or(0);
                 let mut fields = indexmap::IndexMap::new();
-                fields.insert("port".to_string(), Value::Float(port as f64));
+                // NUM §4: a port is an `Int`.
+                fields.insert("port".to_string(), Value::Int(i64::from(port)));
                 let handle = self.register_resource(
                     NativeKind::TcpListener,
                     fields,
